// frontend/src/types/csaf.ts
export type AudienceOfNote = string;
/**
 * Contains the information of what kind of note this is.
 */
export type NoteCategory =
  | "description"
  | "details"
  | "faq"
  | "general"
  | "legal_disclaimer"
  | "other"
  | "summary";
/**
 * Holds the content of the note. Content varies depending on type.
 */
export type NoteContent = string;
/**
 * Provides a concise description of what is contained in the text of the note.
 */
export type TitleOfNote = string;
// ...
export type CategoryOfReference = "external" | "self";
/**
 * Indicates what this reference refers to.
 */
export type SummaryOfTheReference = string;
/**
 * Provides the URL for the reference.
 */
export type URLOfReference = string;
// ...
export interface Note {
  audience?: AudienceOfNote;
  category: NoteCategory;
  text: NoteContent;
  title?: TitleOfNote;
  [k: string]: unknown;
}
// ...
export interface Reference {
  category?: CategoryOfReference;
  summary: SummaryOfTheReference;
  url: URLOfReference;
  [k: string]: unknown;
}